Put little dogs and kittens into Gucci bag
This creates a matching object and puts that into the bag, moving
monster from level into limbo.

If the animal carries amulet of yendor or invocation item, that will
drop on the floor. Any other carried item with disappear from the game
until the pet is back out of the bag.

Can only put one pet into a bag.